Make class fields a special kind of property (#2548)
BREAKING CHANGE: Element kinds FIELD and FIELD_PROTOTYPE and elements Field and FieldPrototype have been removed. Instead, Property inherit previous Field functionality, indicated by property.isField.

+++ tests/compiler/function-expression.ts
@@ -87,7 +87,7 @@ export function semanticallyAnonymous(): void {
   function fnDecl(val: i32): i32 {
     return val;
   }
-  const exprDecl = function fnDecl(val: i64): i64 { // must not shadow
+  const exprDecl = function fnDecl(val: i32): i32 { // must not shadow
     return val;
   };
   assert(fnDecl != exprDecl);
